Pass tests

+++ franken/backbones/wrappers/mace_wrap.py
@@ -398,7 +398,9 @@ class FrankenMACE(torch.nn.Module):
             natoms=partial_config.natoms,
             unit_shifts=torch.from_numpy(unit_shifts).to(dtype=dtype, device=device),
             shifts=torch.from_numpy(shifts).to(dtype=dtype, device=device),
-            edge_index=torch.from_numpy(edge_index).to(device=device).transpose(0, 1),
+            edge_index=torch.from_numpy(edge_index)
+                            .to(dtype=torch.int64, device=device)
+                            .transpose(0, 1),
         )
 
     @staticmethod
